Make updates to the entity details api based on the new schema

+++ app/src/main/java/com/babylon/wallet/android/domain/model/OwnedNonFungibleToken.kt
@@ -1,10 +1,8 @@
 package com.babylon.wallet.android.domain.model
 
-import java.math.BigDecimal
-
 data class OwnedNonFungibleToken(
     val owner: AccountAddress,
-    val amount: BigDecimal,
+    val amount: Long,
     val tokenResourceAddress: String,
     val token: NonFungibleToken? = null
 )
